+++ packages/components/checkbox/src/Checkbox.variants.tsx
@@ -1,15 +1,11 @@
 import { cva, VariantProps } from 'class-variance-authority'
 
-const defaultVariants = {
-  intent: 'primary',
-} as const
-
 export const styles = cva(
   [
-    'peer h-sz-20 w-sz-20 rounded-sm border-md bg-transparent items-center justify-center',
+    'h-sz-20 w-sz-20 border-md peer items-center justify-center rounded-sm bg-transparent',
     'spark-disabled:opacity-dim-3 spark-disabled:cursor-not-allowed spark-disabled:hover:ring-0',
-    'focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-outline-high',
-    'hover:outline-none hover:ring-2 hover:border-primary-container',
+    'focus-visible:ring-outline-high focus-visible:outline-none focus-visible:ring-2',
+    'hover:border-primary-container hover:outline-none hover:ring-2',
   ],
   {
     variants: {
@@ -32,7 +28,9 @@ export const styles = cva(
         ],
       },
     },
-    defaultVariants,
+    defaultVariants: {
+      intent: 'primary',
+    },
   }
 )