+++ packages/utils/internal-utils/src/tw/index.ts
@@ -1,0 +1,31 @@
+/**
+ * A utility function that serves as a workaround for Tailwind CSS IntelliSense issues
+ * in large objects or when regex breaks in VSCode.
+ *
+ * For more context, see: https://github.com/joe-bell/cva/discussions/85#discussioncomment-4568738
+ *
+ * @example
+ *
+ * // before
+ * const styles: cva(['flex bg-primary', 'cursor-pointer'], {
+ *   variants: {
+ *     size: {
+ *       small: "h-sz-24",
+ *       medium: "h-sz-48"
+ *     }
+ *   }
+ * })
+ *
+ * // after
+ * const styles = cva(tw(['flex bg-primary', 'cursor-pointer']), {
+ *   variants: {
+ *     size: {
+ *       small: tw("h-sz-24"),
+ *       medium: tw("h-sz-48"),
+ *     }
+ *   }
+ * })
+ */
+export function tw<T>(v: T) {
+  return v
+}