+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.4](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.3.3...@spark-ui/icons@1.3.4) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/icons
+
+## [1.3.3](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.3.2...@spark-ui/icons@1.3.3) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/icons
+
+## [1.3.2](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.3.1...@spark-ui/icons@1.3.2) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/icons
+
+## [1.3.1](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.3.0...@spark-ui/icons@1.3.1) (2023-04-04)
+
+**Note:** Version bump only for package @spark-ui/icons
+
 # [1.3.0](https://github.com/adevinta/spark/compare/@spark-ui/icons@1.2.2...@spark-ui/icons@1.3.0) (2023-04-03)
 
 ### Bug Fixes