+++ packages/components/radio/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.2](https://github.com/adevinta/spark/compare/@spark-ui/radio@1.3.1...@spark-ui/radio@1.3.2) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/radio
+
+## [1.3.1](https://github.com/adevinta/spark/compare/@spark-ui/radio@1.3.0...@spark-ui/radio@1.3.1) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/radio
+
 # [1.3.0](https://github.com/adevinta/spark/compare/@spark-ui/radio@1.2.2...@spark-ui/radio@1.3.0) (2023-04-03)
 
 ### Features