+++ packages/components/slot/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/adevinta/spark/compare/@spark-ui/slot@1.2.2...@spark-ui/slot@1.2.3) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/slot
+
 ## [1.2.2](https://github.com/adevinta/spark/compare/@spark-ui/slot@1.2.1...@spark-ui/slot@1.2.2) (2023-03-31)
 
 **Note:** Version bump only for package @spark-ui/slot