+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.6](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.2.5...@spark-ui/switch@1.2.6) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.2.5](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.2.4...@spark-ui/switch@1.2.5) (2023-04-05)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.2.4](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.2.3...@spark-ui/switch@1.2.4) (2023-04-04)
+
+**Note:** Version bump only for package @spark-ui/switch
+
 ## [1.2.3](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.2.2...@spark-ui/switch@1.2.3) (2023-04-03)
 
 ### Bug Fixes